clean fix size, prepare to implement selector for notes
